src/interface/slack/requirements.txt
Flask==2.0.3
slack_sdk==3.12.0
python-dotenv==0.19.2
requests==2.26.0
asyncio==3.4.3